+++ matrix-sdk-android/src/debug/java/org/matrix/android/sdk/internal/network/interceptors/FormattedJsonHttpLogger.kt
@@ -17,7 +17,6 @@
 package org.matrix.android.sdk.internal.network.interceptors
 
 import androidx.annotation.NonNull
-import org.matrix.android.sdk.BuildConfig
 import okhttp3.logging.HttpLoggingInterceptor
 import org.json.JSONArray
 import org.json.JSONException
@@ -38,31 +37,28 @@ class FormattedJsonHttpLogger : HttpLoggingInterceptor.Logger {
      */
     @Synchronized
     override fun log(@NonNull message: String) {
-        // In RELEASE there is no log, but for sure, test again BuildConfig.DEBUG
-        if (BuildConfig.DEBUG) {
-            Timber.v(message)
+        Timber.v(message)
 
-            if (message.startsWith("{")) {
-                // JSON Detected
-                try {
-                    val o = JSONObject(message)
-                    logJson(o.toString(INDENT_SPACE))
-                } catch (e: JSONException) {
-                    // Finally this is not a JSON string...
-                    Timber.e(e)
-                }
-            } else if (message.startsWith("[")) {
-                // JSON Array detected
-                try {
-                    val o = JSONArray(message)
-                    logJson(o.toString(INDENT_SPACE))
-                } catch (e: JSONException) {
-                    // Finally not JSON...
-                    Timber.e(e)
-                }
+        if (message.startsWith("{")) {
+            // JSON Detected
+            try {
+                val o = JSONObject(message)
+                logJson(o.toString(INDENT_SPACE))
+            } catch (e: JSONException) {
+                // Finally this is not a JSON string...
+                Timber.e(e)
             }
-            // Else not a json string to log
+        } else if (message.startsWith("[")) {
+            // JSON Array detected
+            try {
+                val o = JSONArray(message)
+                logJson(o.toString(INDENT_SPACE))
+            } catch (e: JSONException) {
+                // Finally not JSON...
+                Timber.e(e)
+            }
         }
+        // Else not a json string to log
     }
 
     private fun logJson(formattedJson: String) {